refactor to a library
