+++ Cargo.toml
@@ -1,9 +1,15 @@
 [package]
-name = "rust-zero2prod"
+name = "email_newsletter"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path="src/main.rs"
+name="email_newsletter"
 
 [dependencies]
 actix-web="4"
