+++ src/lib.rs
@@ -1,0 +1,15 @@
+use actix_web::{web, App, HttpResponse, HttpServer, Responder};
+
+async fn health_check() -> impl Responder {
+    HttpResponse::Ok()
+}
+
+pub async fn run() -> std::io::Result<()> {
+    HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+    })
+    .bind(("127.0.0.1", 8080))?
+    .run()
+    .await
+}
